requirements.txt
amqp==5.3.1
asgiref==3.9.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
limits==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
ordered-set==4.1.0
packaging==25.0
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
rich==13.9.4
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
